Properly fix draggable
With the introduction of reordering draggable windows, they can
be replaced between pointerdown & pointermove. Since the target has
changed, the events that were meant to be captured won't send
updates to the newly inserted element.

Fix is to have a global draggable listener for listening to events
where it will not be reshuffled/reordered.